Refactor de_export, extract DEModel
Move `de_export.DEModel` to `de_model`.

Closes #2306. de_export is now down to 1.5K lines, from previously 4K+.
Some further cleanup would be good, but it's a start.
